mock-code: add 'ignore_paths' argument
The `ignore_files` parameter of the mock-code fixture allows ignoring
files only based on their filename, but not based on their path in the
the run directory

This does not cover the use case when a code creates a subfolder
structure and one would like to ignore all files in certain subfolders.

This PR adds a `ignore_paths` option to cover that use case (and
deprecates the `ignore_files` option).